Arduino and QA Testing
I decided to build something quick and simple using my arduino at home and try out some simple unit test scripts. Hopefully, this would be interesting enough for the team.

The attached files are the image of the circuit board I created, the codes and the results. It is a simple temperature sensor with a LCD display (followed the simple tutorial https://create.arduino.cc/projecthub/TheGadgetBoy/making-lcd-thermometer-with-arduino-and-lm35-36-c058f0).

In order to test out whether the board is working properly before the run, we can run a simple unit test to find out. I attached the codes. The first one is testing some simple function using assertion included the continuous test. The second is a separated unit test for only the continuous test.

By comparing the results of the two tests, one can find out continuous test failed in the first case but passed in the second one. This can be potentially explained by the fact that other tests are simultaneously running which affects the testing time.